backend/requirements.txt
alembic==1.17.2
APScheduler==3.10.4
certifi==2025.11.12
charset-normalizer==3.4.4
fastapi==0.125.0
greenlet==3.3.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.1
uvicorn==0.38.0